DB1000N | Add support for the restart option after the auto-update for the Windows OS
