+++ src/utils/ota/shared_test.go
@@ -1,6 +1,3 @@
-//go:build linux
-// +build linux
-
 package ota
 
 import (
